Size_t is not an uint64_t, size_t is not an uint64_t, size_t is not an uint64_t

+++ lib/Logger/Logger.h
@@ -307,7 +307,7 @@ class Logger {
   }
 
   static void initialize(application_features::ApplicationServer&, bool,
-                         size_t maxQueuedLogMessages);
+                         uint64_t maxQueuedLogMessages);
   static void shutdown();
   static void flush() noexcept;
